Rework MapTiles generation around generic map and tile types
[MOD] IGenerate: replaced IGenerate<T> with IGenerate<MAP_SIZE_TYPE, TILE_TYPE, TILE_POS_UNIT, TILE_SIZE_UNIT>. Added MapSize, TileSize, CreateTile(); dropped NextMap/RandomMap/DailyMap/LoadMap/GenerateMap from non-generic base.
[MOD] BaseMap: split into non-generic base holding Tileset, Generator and static GetCoordinatesInRadius(), plus generic BaseMap<T, TU, TS> holding Tiles dictionary. Tile search methods now typed.
[ADD] Map2DTiles. Concrete BaseMap<Tile2D, Vector2Int, int> with surrounding and encompassing tile lookups, radius and square coordinate helpers, FindAllSimilarConnected().
[MOD] Default2DGenerator: typed to Tile2D, takes tileSize, folded mapSeed field into Seed property.
[MOD] RadialGenerator: typed to Tile2D, takes tileSize, parent and a GetCoordinatesInRadius delegate. Start/exit placement and second-pass tile rules commented out.
[MOD] Tile2D: tile size unit float -> int.
[MOD] MapTester: added tileSize field, lowered mapRadius minimum 10 -> 2, typed tile dictionary to Tile2D.
[ADD] GrassTilePrefab, RockTilePrefab, SandTilePrefab, WaterTilePrefab.
[MOD] Tileset: replaced placeholder Default entry with Grass, Rock, Sand and Water, each with spawn weight and tile variant.
[MOD] MapTileGenerationSample scene: generated tile instances saved from MapTester run.

+++ Samples~/MapTiles/Scripts/Tiles/Tile2D.cs
@@ -2,11 +2,11 @@
 
 namespace Tiles
 {
-    public class Tile2D : BaseTile<Vector2Int, float>
+    public class Tile2D : BaseTile<Vector2Int, int>
     {
         public override Vector2Int Position { get; set; }
         
-        public override float Size => 2f;
+        public override int Size => 2;
 
         protected override void OnClickDown()
         {
